Add KHLs from packetstream to cache
 G2Packet.c
	forward info gathered from /KHL/CH and /KHL/NH to cache

 lib/combo_addr.h
	add a hashing function for a combo_addr
 G2KHL.h
	add a parameter if the node is in the local cluster (??? what ever that is ;)
 G2KHL.c
	use it, minor adjustments

+++ G2Connection.h
@@ -83,7 +83,7 @@ typedef struct g2_connection
 
 	/* Internal States */
 	union combo_addr sent_addr;
-	double           time_diff;
+	long             time_diff;
 	struct
 	{
 		time_t        PI;
